requirements.txt
flask==3.0.0
python-dotenv==1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.51
pydantic>=2.5.2
langsmith>=0.1.0
supabase>=2.3.4
openai>=1.0.0
tiktoken>=0.5.2 